+++ src/middleware/mod.rs
@@ -10,10 +10,12 @@ mod logger;
 #[cfg(feature = "session")]
 mod session;
 mod defaultheaders;
+mod errhandlers;
 pub mod cors;
 pub mod csrf;
 pub use self::logger::Logger;
-pub use self::defaultheaders::{DefaultHeaders, DefaultHeadersBuilder};
+pub use self::errhandlers::ErrorHandlers;
+pub use self::defaultheaders::DefaultHeaders;
 
 #[cfg(feature = "session")]
 pub use self::session::{RequestSession, Session, SessionImpl, SessionBackend, SessionStorage,